Modularization

+++ src/main/webapp/controller/deletebookcontroller.jsp
@@ -3,6 +3,7 @@
 <%@page import="com.liberarymanagement.daoimpl.AdminDaoImpl"%>
 <%@page import="com.liberarymanagement.dao.AdminDao"%>			
 <% 
+if(session.getAttribute("adminemail")!=null){
       String name =request.getParameter("name");
 	AdminDao daoImpl = new AdminDaoImpl();
 		if(daoImpl.deleteBook(name)){
@@ -11,4 +12,8 @@
 		else{
 			response.sendRedirect("deletebook.jsp?msg=deleteFail");
 		}	
+	}
+	else{
+		response.sendRedirect("index.jsp?msg=loginfirst");
+	}
 %>
